4.2

+++ CUDA_TEST4/LaserRange.h
@@ -44,7 +44,6 @@ private:
 	const double offset;//offset constant
 	const double h_cm;//distance between center of camera and laser
 	unsigned int pixel_from_center;//brightest pixel location from center
-
 	void Preprocess(void *img, IplImage *imgTemp);
 };
 
@@ -76,20 +75,19 @@ typedef struct MPoint
 
 typedef struct MatrixUnion
 {
-	double **X;
-	double **XT;
-	double *Z;
-	double *B;
-	double **SA;
+	long long int **X;
+	long long int **XT;
+	long long int *Z;
+	long long int *B;
+	long long int **SA;
 	double **SAN;//逆矩阵
 	double **SC;
 	double **in_v;//逆矩阵的转置
-	double **BC; // 
+	double **BC; 
 	double **L; //LU过程矩阵L
 	double **U; //LU过程矩阵U
-	double **A_mirror;  //原矩阵的复制
+	long long int **A_mirror;  //原矩阵的复制
 	int *P; //LUP 记录列坐标
-
 }MatrixUnion;
 
 //张正友相机标定法
